Gutted create_scene_xml process, added bvxm/algo/bvxm_create_scene_xml.
