+++ contrib/brl/bseg/bvxm/pro/bvxm_processes.h
@@ -64,8 +64,6 @@ DECLARE_FUNC_CONS(bvxm_dem_to_xyz_process);
 DECLARE_FUNC_CONS(bvxm_dem_to_xyz_process2);
 
 //Basic operation
-DECLARE_FUNC_CONS(bvxm_create_scene_xml_process);
-DECLARE_FUNC_CONS(bvxm_create_scene_xml_large_scale_process);
 DECLARE_FUNC_CONS(bvxm_scene_model_dir_process);
 DECLARE_FUNC_CONS(bvxm_scene_origin_process);
 DECLARE_FUNC_CONS(bvxm_scene_lvcs_process);
